moving kofola source to utils
