+++ src/complement/decomposer.hpp
@@ -12,7 +12,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "../kofola.hpp"
+#include "../util/kofola.hpp"
 
 #include <functional>
 #include <set>
