+++ src/main.cpp
@@ -16,7 +16,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 // kofola
-#include "kofola.hpp"
+#include "util/kofola.hpp"
 #include "complement/complement_tela.hpp"
 #include "inclusion/emptiness_check.hpp"
 #include "util/util.hpp"
