+++ src/util/kofola.cpp
